Refactor and fix various code issues across multiple files
- Updated string formatting in wolgmode.cpp to ensure proper type casting.
- Improved debug message formatting in Mission05.cpp for better precision.
- Defined INVALID_MILES_PROVIDER in AudibleSound.h for clarity.
- Corrected initialization of m_hFilter in FilteredSound.cpp and WWAudio.cpp to use INVALID_MILES_PROVIDER.
- Fixed string handling in animatedsoundmgr.cpp and other files by using Peek_Buffer().
- Changed operator overload definitions in several classes to remove redundant class names.
- Added const qualifiers to getter methods in various classes for better const-correctness.
- Updated texture loading code to use uint8 for palette arrays in textureloader.cpp and related files.
- Enhanced collision handling in physics-related files to use static_cast for clarity.
- Removed unnecessary newlines and improved formatting in several header files.

+++ Code/ww3d2/part_buf.cpp
@@ -2540,7 +2540,7 @@ void ParticleBufferClass::Update_Visual_Particle_State(void)
 				RandomRotationEntries[part & NumRandomRotationEntriesMinus1] * (float)part_age +
 				RandomOrientationEntries[part & NumRandomOrientationEntriesMinus1];
 			
-			orientation[part] = (uint)(((int)(tmp_orient * 256.0f)) & 0xFF);
+			orientation[part] = static_cast<uint8>(((int)(tmp_orient * 256.0f)) & 0xFF);
 		}
 
 		// Ensure the current frame keyframe is correct, and calculate frame state
@@ -2555,7 +2555,7 @@ void ParticleBufferClass::Update_Visual_Particle_State(void)
 				FrameKeyFrameDeltas[fkey] * (float)(part_age - FrameKeyFrameTimes[fkey]) +
 				RandomFrameEntries[part & NumRandomFrameEntriesMinus1];
 			
-			frame[part] = (uint)(((int)(tmp_frame)) & 0xFF);
+			frame[part] = static_cast<uint8>(((int)(tmp_frame)) & 0xFF);
 		}
 
 		// Ensure the current frame keyframe is correct, and calculate frame state
@@ -2640,7 +2640,7 @@ void ParticleBufferClass::Update_Visual_Particle_State(void)
 				RandomRotationEntries[part & NumRandomRotationEntriesMinus1] * (float)part_age +
 				RandomOrientationEntries[part & NumRandomOrientationEntriesMinus1];
 			
-			orientation[part] = (uint)(((int)(tmp_orient * 256.0f)) & 0xFF);
+			orientation[part] = static_cast<uint8>(((int)(tmp_orient * 256.0f)) & 0xFF);
 		}
 
 		// Ensure the current frame keyframe is correct, and calculate frame state
@@ -2655,7 +2655,7 @@ void ParticleBufferClass::Update_Visual_Particle_State(void)
 				FrameKeyFrameDeltas[fkey] * (float)(part_age - FrameKeyFrameTimes[fkey]) +
 				RandomFrameEntries[part & NumRandomFrameEntriesMinus1];
 			
-			frame[part] = (uint)(((int)(tmp_frame)) & 0xFF);
+			frame[part] = static_cast<uint8>(((int)(tmp_frame)) & 0xFF);
 		}
 
 		// Ensure the current frame keyframe is correct, and calculate frame state
